+++ release.py
@@ -8,13 +8,18 @@ from __future__ import (absolute_import, division, print_function)
 __metaclass__ = type
 
 import argparse
+import base64
 import os
 import os.path
+import re
 import shutil
+import subprocess
 import tarfile
+import tempfile
 
 from utils import (
     argcomplete,
+    get_version,
     OMI_REPO,
 )
 
@@ -23,35 +28,63 @@ def main():
     """Main program body."""
     args = parse_args()
 
+    version = get_version()
+
     if args.print_tag:
-        with open(os.path.join(OMI_REPO, 'omi.version'), mode='r') as fd:
-            version_lines = fd.read().splitlines()
+        print("v%s.%s.%s-pwsh" % (version.major, version.minor, version.patch))
 
-        version_info = {}
-        for line in version_lines:
-            k, v = line.split('=', 1)
-            version_info[k] = v
+    elif args.output_dir:
+        if os.path.exists(args.output_dir):
+            shutil.rmtree(args.output_dir)
+        os.makedirs(args.output_dir)
 
-        print("v%s.%s.%s-pwsh" % (version_info['OMI_BUILDVERSION_MAJOR'], version_info['OMI_BUILDVERSION_MINOR'],
-                                  version_info['OMI_BUILDVERSION_PATCH']))
-
-    elif args.pipeline_artifacts:
-        # This renames the Azure DevOps pipeline files to the format desired for a GitHub release asset
-        for distribution in os.listdir(args.pipeline_artifacts):
-            artifact_dir = os.path.join(args.pipeline_artifacts, distribution)
+        # Create a tar.gz for each distribution libs for the GitHub release
+        lib_path = os.path.join(OMI_REPO, 'PSWSMan', 'lib')
+        for distribution in os.listdir(lib_path):
+            artifact_dir = os.path.join(lib_path, distribution)
+            artifact_tar = os.path.join(args.output_dir, distribution) + '.tar.gz'
 
             if distribution.startswith('.') or not os.path.isdir(artifact_dir):
                 continue
 
-            print("Creating '%s.tar.gz'" % artifact_dir)
-            with tarfile.open('%s.tar.gz' % artifact_dir, 'w:gz') as tar:
+            print("Creating '%s'" % artifact_tar)
+            with tarfile.open(artifact_tar, 'w:gz') as tar:
                 for lib_name in os.listdir(artifact_dir):
                     if lib_name == '.':
                         continue
                     print("\tAdding '%s' to tar" % lib_name)
                     tar.add(os.path.join(artifact_dir, lib_name), arcname=lib_name)
 
-            shutil.rmtree(artifact_dir)
+        # Create the PSWSMan nupkg
+        pwsh_command = u'''$ErrorActionPreference = 'Stop'
+
+$outputDir = '%s'
+$repoParams = @{
+  Name = 'PSWSManRepo'
+  SourceLocation = $outputDir
+  PublishLocation = $outputDir
+  InstallationPolicy = 'Trusted'
+}
+if (Get-PSRepository -Name $repoParams.Name -ErrorAction SilentlyContinue) {
+    Unregister-PSRepository -Name $repoParams.Name
+}
+Register-PSRepository @repoParams
+
+try {
+    Publish-Module -Path ./PSWSMan -Repository $repoParams.Name
+} finally {
+    Unregister-PSRepository -Name $repoParams.Name
+}
+''' % args.output_dir
+        encoded_pwsh = base64.b64encode(pwsh_command.encode('utf-16-le')).decode()
+
+        print("Creating PSWSMan nupkg")
+        subprocess.check_call(['pwsh', '-EncodedCommand', encoded_pwsh], cwd=OMI_REPO)
+
+        nupkg_name = None
+        for name in os.listdir(args.output_dir):
+            if name.endswith('.nupkg'):
+                print("Published PSWSMan to '%s'" % os.path.join(args.output_dir, name))
 
 
 def parse_args():
@@ -65,18 +98,18 @@ def parse_args():
                            action='store_true',
                            help='Print the tag number for the release.')
 
-    run_group.add_argument('--pipeline-artifacts',
-                           dest='pipeline_artifacts',
+    run_group.add_argument('--output-dir',
+                           dest='output_dir',
                            action='store',
-                           help='The Azure DevOps pipeline artifact store directory to process.')
+                           help='The directory to create the release artifacts at.')
 
     if argcomplete:
         argcomplete.autocomplete(parser)
 
     args = parser.parse_args()
 
-    if not args.print_tag and not args.pipeline_artifacts:
-        parser.error('argument --print-tag or --pipeline-artifacts must be seet')
+    if not args.print_tag and not args.output_dir:
+        parser.error('argument --print-tag or --output-dir must be seet')
 
     return args