+++ CHANGELOG.md
@@ -11,6 +11,7 @@ It documents the changes in each of the tagged releases
 
 ### Changes
 
++ Created `PSWSMan` which is a PowerShell module uploaded to the PowerShell Gallery that can install and manage the OMI libraries for you
 + Build `libpsrpclient` as well and add it to the release artifacts
 
 ## 1.2.1 - 2020-09-26